Another helper script to deal with GMAP errors
Script removes lines from the BLAT gff3 file that contain hits to sequences that have been removed due to "problem sequence" errors. Reason this script was made is because BLAT was run on the whole transcriptomes prior to fixing this error with GMAP, and I don't want to re-run BLAT from scratch.