Calculate percentage of protein that is LCR
Helper script to manually curate the output gene models. From more hand-on manual inspection, most gene models that are overlapped by repeat masked sequence that I think should be removed from the final annotation have a high proportion of low complexity regions. This script, in tandem with the repeat masked proportion, should make it easier to automatically remove sequences which meet both criteria strongly, and to manually check sequences that fall into a 'borderline' region wherein I'm not sure if they should be given a chance to be considered genuine or discarded as a crappy gene model.