Modify CDS regions of gff3 file
Helper script originally intended to find the starting frame of the CDS region and modify this in the file itself. Later on realised that doing this wouldn't properly represent the genomic information, so I decided to modify the gff3_to_fasta.py script. This script may be useful later, or might provide a foundation for reworking to be something else.