Assess evidence for gene models
Combines evidence from multiple sources to find gene models that are likely to be false. From my testing, even with a wide range of evidence, it's hard to be confident that the identified gene models are actually false, even for sequences with >90% LCR since many have seemingly good BLAST hits e.g., to histone acetyltransferases, and so using a too-strict E-value cut-off will result in the removal of genuine gene models as well. Thus, I likely won't be using this script but it might be useful for another purpose later.