WIP: Final stages
My goal is for this script to "patch" a genome sequence using RNAseq transcripts to correct indels using sequence that is biologically accurate. You can theoretically "fix" gene models by introducing N's or deleting bases to offset frameshifts, but it can be difficult to figure out where the correct location is to place these modifications. Using transcripts from the species itself, these errors should be minimised even if the result is a bit of a mosaic of individuals.