WIP: Working with some minor issue
Currently I can't figure out how to get the RNAseq "patch" for all regions that don't have exact matches. Exact matches make it easy, but inexact matches mean that we can't reliably get the start bit. Because of this, I'm thinking of changing the code's function a little to use MUSCLE alignment of the "patch" against the joining region to 1) find the borders of the RNAseq patch reliably, and 2) identify the locations of the indels to just remove these indels rather than full-scale "patching" the RNAseq region into the genome, since this might introduce variants from other individuals into the genome. While this isn't bad imo (the original human genome was a mosaic) it's probably easier to convince others that this program isn't doing anything "weird" if I keep its interaction with the genome sequence to an absolute minimum.