WIP: Simplifying program
From testing with an unnecessarily more complex version of this program, I realised that it was quite powerful and could be simplified to work in a more general manner when allowing non-exact matches at the borders of exons. This naturally corrected all of the scenarios I was trying to handle (gene model fragmentation across an exon, within-exon fragmentation) and should also naturally handle most occurrences of gene model fragmentation across an intron. It's possible I'll generalise the program even more to just work directly with the GMAP alignments - it might be a bit risky, but I think with some safe guards it'll help us find even more gene models.