Extend table with exon,intron details
Adds four new columns to the annotation table. These include 1) transcript and CDS lengths, 2) transcript and cds exon regions (gives coordinates from 1->len(sequence) with respect to whether it's a transcript or CDS), 3) exon support from transcripts (behaves similarly to gff3_statistics.py in the deGRIT repository), and 4) intron lengths. Combined, this information makes it easier to manually inspect sequences and make relevant judgements. Also - this marks the likely end to these annotation extension scripts. In the future I may decide to include more information in this table, or to bundle these scripts to automate the process, but the day I do that is not today.