Add trim function
Function to trim bits from the start and/or end of sequences in fasta or fastq format. Added in a few general functions for handling fastq files that don't conform to Biopython's expectations (required for a file I was trying to work with) and for generating temporary files as part of this process.