Multiple inputs, main loop finished, needs final checks
Script modified to accept multiple GFF3 files on the command line. This is so we can combine their outputs within the one script. The main loop is finished now, it works for handling most gene models well. One time an assert statement wasn't met, but when I re-ran it it worked. Bug in one of the modules? Unsure, but there is nothing random in my own code so I must assume that. There are a few scenarios that likely need specific curation checks to fix such as one scenario where there is a microintron that, when filled in, results in a single exon gene. Removing genes that solely consist of microexons should resolve this problem.