Not good enough; shelving for now
The idea is still valid, but it's too difficult to effectively make the kind of fixes I was planning (i.e., automatic detection and fixing of merged gene models). In Aulactinia there's every chance I can fix about 100 gene models, but I am unable to be certain what he error rate of this program will be. Since it is affecting so few genes and I can't be sure about whether it's truly working well or not, I'm going to shelve it. I might still make a CDS location fixer especially since Actinia seems like it might need it, but I might just leave it as is. PASA should have done a good enough job (hopefully) and BUSCO scores are still reasonable even for the CDS extracted regions. Bit disappointing that this didn't work out, but it happens.