Disallow fragment transcripts becoming models
Extra stringency check reduced the number of models in Aulactinia by about 100 which seems reasonable. The check is quite simple, if the transcript doesn't end with a stop codon we ignore it. Previously, it was possible that we'd build a gene model that is quite likely to be fragmented. There is the possibility that the stop codon correction would naturally fix the model, but if an entire exon was missing from the tail of a transcript, our new model would be flawed.