GGF behaviour change & merge info
GGF's behaviour was noticed to be a bit different after changing SSW handling, and for the worse in some cases (BUSCO scores reduced). I looked into the reason and found that my splice filtration system was killing a few genes that were obviously better than others due to their length being longer and including 1 noncanonical splice. Now, we only filter on splice sites if there is a large difference between the percentage values (in the 3 cases I observed, the % difference was about 5%, so 10% should be a safe cutoff).