Final splice rules filtering modification
I did some actual testing with this before pushing the commit for once. I'm happy with the results this produces. Because we have a bad splice remover later in the program, having length take priority of splice rules should be fine, and it so we don't accept a fragment of a gene (which is in a region with only perfet splices) over the larger gene (which has a few non-canonical splices but is still good).