Alternative to gffcompare
gffcompare doesn't seem to work well with S. cerevisiae's annotation. I'm unsure the reasons why, but mikado works better overall. It throws up a few exceptions which don't appear to halt program execution which is slightly concerning, but it's the better of the two options from what I can see.