WIP: about to change index_contig_reorder
Code is starting to become functional, but it's clear that there are limitations to this. MITOS2 doesn't provide ideal annotations, and the start position of the sequence can influence the annotation's correctness. Additionally, it's crappy GFF2 formatting means you can't tell which intron is actually the start of the gene. Frustrating.