EGF compare_novels_store_rejects changes
I encountered a situation that I can only assume indicates some unintended behaviour in this code. It is possible, in a case with 3 genes, for 2 genes to be non-overlapping and compatible, and for 1 of these said genes to be overruled by a 3rd one which does not overrule the other gene. Essentially, the "poison" gene removes another gene before itself being removed, thus resulting in only 1 gene making it through rather than the "poisoned" gene and the other one which would make for 2 genes. I need to test this in practice to see what changes it brings about, however.