EGF2 seg change and final cov cutoff
Seg is still having issues with spurious dropping of otherwise good sequences that just happen to consist primarily of domains that are somewhat low-complexity/repetitive. Otherwise, the final covcutoff has been put in because we're probably annotating a lot of sequences that simply aren't similar to the original exonerate query and therefore can't reliably be considered a toxin candidate.